p3/chronos-forecast/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
torch==2.1.0
git+https://github.com/amazon-science/chronos-forecasting.git
google-generativeai==0.3.2